% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{active_cycle_check}
\alias{active_cycle_check}
\title{Checks if a graph contains active cycles}
\usage{
active_cycle_check(DAG)
}
\arguments{
\item{DAG}{Directed Acyclic Graph}
}
\value{
a list containing a boolean specifying if DAG contains active cycles,
number of active cycles, and list of the active cycles.
}
\description{
Checks if a graph contains active cycles
}
