% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hill_climbing.R
\name{allowed.operations.general}
\alias{allowed.operations.general}
\title{Finds all operations resulting in a restricted DAG}
\usage{
allowed.operations.general(DAG)
}
\arguments{
\item{DAG}{the current DAG}
}
\value{
a matrix with 3 columns `from`, `to`, `operation`.
Possible operations are "set", "drop" and "reverse".
}
\description{
Finds all operations resulting in a restricted DAG
}
