% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaAssignement.R
\name{copula_is_specified}
\alias{copula_is_specified}
\title{Checks if a copula is specified}
\usage{
copula_is_specified(DAG, order_hash, w, v, cond)
}
\arguments{
\item{DAG}{Directed Acyclic Graph}

\item{order_hash}{hashmap of orders}

\item{w}{node in DAG}

\item{v}{node in DAG}

\item{cond}{vector of nodes in DAG}
}
\value{
TRUE if d-sep(w, v | cond)
}
\description{
Checks if a copula is specified
}
