% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_PCBN.R
\name{create_copula_tag}
\alias{create_copula_tag}
\title{Creates copula tags}
\usage{
create_copula_tag(DAG, order_hash, w, v, cond_set = NULL)
}
\description{
Creates copula tags
}
\examples{

DAG = create_DAG(3)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')

order_hash = r2r::hashmap()
order_hash[['U3']] = c("U1", "U2")

create_copula_tag(DAG = DAG, order_hash = order_hash,
                  w = "U1", v = "U2", cond_set=NULL)

create_copula_tag(DAG = DAG, order_hash = order_hash,
                  w = "U2", v = "U3", cond_set="U1")

}
