% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{create_DAG}
\alias{create_DAG}
\title{Create empty graph}
\usage{
create_DAG(N_nodes)
}
\arguments{
\item{N_nodes}{An integer equal to the number of nodes}
}
\value{
A bnlearn graph object with `N_nodes` nodes and no arcs
}
\description{
`create_DAG()` creates a directed graph with a total of `N_nodes` nodes and no arcs.
The graph is a bnlearn object.
The nodes are named `U1`, `U2`, etc.
}
\examples{
create_DAG(6)
create_DAG(10)
}
\seealso{
[bnlearn::empty.graph()] which this function wraps.
}
