% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{create_empty_DAG}
\alias{create_empty_DAG}
\title{Create empty DAG}
\usage{
create_empty_DAG(N_nodes)
}
\arguments{
\item{N_nodes}{An integer equal to the number of nodes}
}
\value{
A \code{bnlearn} graph object with `N_nodes` nodes and no arcs
}
\description{
This function creates a directed graph with a total of `N_nodes` nodes and
no arcs. The nodes are named `U1`, `U2`, etc.
}
\examples{
create_empty_DAG(6)
create_empty_DAG(10)

}
\seealso{
[bnlearn::empty.graph()] which this function wraps.
}
