% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_PCBN.R
\name{create_margin_tag}
\alias{create_margin_tag}
\title{Creates a tag for a (conditional) margin}
\usage{
create_margin_tag(DAG, order_hash, v, cond_set)
}
\description{
For a conditional margin u_{v|K} the key cannot only be of the form "v|K".
We also require that the string represents the recursion of h-functions
with which the margin has been computed.
}
\examples{

DAG = create_DAG(3)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')

order_hash = r2r::hashmap()
order_hash[['U3']] = c("U1", "U2")

create_margin_tag(DAG = DAG, order_hash = order_hash,
                  v = "U2", cond_set=NULL)

create_margin_tag(DAG = DAG, order_hash = order_hash,
                  v = "U3", cond_set="U1")

}
