% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaAssignement.R
\name{extend_orders}
\alias{extend_orders}
\title{Fills in all possible orders for the next node for each possible order}
\usage{
extend_orders(DAG, all_orders, node)
}
\arguments{
\item{DAG}{Directed Acyclic Graph}

\item{all_orders}{list of orders}

\item{node}{node}
}
\value{
list of order hashmaps
}
\description{
Fills in all possible orders for the next node for each possible order
}
