% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaAssignement.R
\name{find_all_orders}
\alias{find_all_orders}
\title{Finds all possible copula assignments given a DAG}
\usage{
find_all_orders(DAG)
}
\arguments{
\item{DAG}{Directed Acyclic Graph}
}
\value{
a list of hashmaps containing the possible orders
}
\description{
Finds all possible copula assignments given a DAG
}
