% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaAssignement.R
\name{find_all_orders_v}
\alias{find_all_orders_v}
\title{Finds all possible orders of node v given previous copula assignments}
\usage{
find_all_orders_v(DAG, v, order_hash)
}
\arguments{
\item{DAG}{Directed Acyclic Graph}

\item{v}{node}

\item{order_hash}{hashmap of orders}
}
\value{
list of vectors containing all possible orders for v
}
\description{
Finds all possible orders of node v given previous copula assignments
}
