% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaAssignement.R
\name{find_B_minus_O}
\alias{find_B_minus_O}
\title{Finds the smallest B-set of v larger than the current partial order}
\usage{
find_B_minus_O(B_sets, partial_order)
}
\arguments{
\item{B_sets}{list of B-sets for a particular node}

\item{partial_order}{order list of parents for particular node}
}
\value{
Smallest B-set strictly larger than the partial order
}
\description{
Finds the smallest B-set of v larger than the current partial order
}
