% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{find_B_sets}
\alias{find_B_sets}
\title{loops over the nodes of the graph, finding all B-sets for each node}
\usage{
find_B_sets(DAG)
}
\arguments{
\item{DAG}{A bnlearn graph object}
}
\value{
A list of B-sets for each node, a boolean specifying
if the graph contains interfering v-structures or not,
and a list containing nodes forming the interfering v-structures.
}
\description{
loops over the nodes of the graph, finding all B-sets for each node
}
