% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{fit_all_orders}
\alias{fit_all_orders}
\title{Fit all possible orders given a DAG}
\usage{
fit_all_orders(data, DAG, familyset = c(1, 3, 4, 5, 6), reuse_hash = FALSE)
}
\arguments{
\item{data}{data frame}

\item{DAG}{Directed Acyclic Graph}

\item{familyset}{vector of copula families}
}
\value{
list containing best fit and all fitted models
}
\description{
Fit all possible orders given a DAG
}
