% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hill_climbing.R
\name{hill.climbing.PCBN}
\alias{hill.climbing.PCBN}
\title{Hill climbing algorithm for restricted PCBNs}
\usage{
hill.climbing.PCBN(
  data,
  familyset = c(1, 3, 4, 5, 6),
  verbose = 2,
  start = NULL,
  e = NULL,
  score_metric = "BIC"
)
}
\arguments{
\item{data}{data frame}

\item{familyset}{vector of copula families}

\item{verbose}{if \code{0}, don't print anything.
If \code{verbose >= 1}, print information about the procedure.}

\item{start}{starting Directed Acyclic Graph. If \code{NULL} (the default),
we start with the empty graph.}

\item{e}{environment containing all the hashmaps}

\item{score_metric}{name of the metric used to choose the best order.
Possible choices are \code{logLik}, \code{AIC} and \code{BIC}.}
}
\value{
DAG which locally maximizes BIC based score function
}
\description{
Hill climbing algorithm for restricted PCBNs
}
\examples{

DAG = create_empty_DAG(4)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U4')
DAG = bnlearn::set.arc(DAG, 'U3', 'U4')

order_hash = r2r::hashmap()
order_hash[['U3']] = c("U2", "U1")
order_hash[['U4']] = c("U2", "U3")

fam = matrix(c(0, 0, 1, 0,
               0, 0, 1, 1,
               0, 0, 0, 1,
               0, 0, 0, 0), byrow = TRUE, ncol = 4)
tau = 0.8 * fam

my_PCBN = new_PCBN(
  DAG, order_hash,
  copula_mat = list(tau = tau, fam = fam))

mydata = PCBN_sim(my_PCBN, N = 5)

result = hill.climbing.PCBN(data = mydata, familyset = 1)

}
