% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaAssignement_possibleCandidates.R
\name{incoming_arc}
\alias{incoming_arc}
\title{Checks if w can be added by an incoming arc}
\usage{
incoming_arc(DAG, w, v, order_v, order_hash)
}
\arguments{
\item{DAG}{Directed Acyclic Graph.}

\item{w, v}{nodes in DAG. \code{w} is assumed to be a parent of \code{v}.}

\item{order_v}{partial order for node v}

\item{order_hash}{hashmap of parental orders}
}
\value{
a node \code{o} such \code{w} is a parent of \code{o},
and \code{w} can be used as an incoming arc to \code{v} by the node \code{o}.
If no such \code{o} can be found, \code{w} cannot be used as a potential
candidate for the order of \code{v} by incoming arc.
This function returns then \code{NULL}.
}
\description{
Checks if w can be added by an incoming arc
}
