% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{increasing_B_set_check}
\alias{increasing_B_set_check}
\title{Checks if the B-sets for a particular node form an increasing sequence.}
\usage{
increasing_B_set_check(B_set)
}
\arguments{
\item{B_set}{list containig nodes in a B-set.}
}
\value{
TRUE if the list forms an ordered sequence, FALSE if not.
}
\description{
Checks if the B-sets for a particular node form an increasing sequence.
}
