% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_B_sets.R
\name{interfering_vstrucs_check}
\alias{interfering_vstrucs_check}
\title{Checks if graph has interfering v-structures}
\usage{
interfering_vstrucs_check(DAG)
}
\arguments{
\item{DAG}{Directed Acyclic Graph}
}
\value{
TRUE if graph contains interfering vs and FALSE if not
}
\description{
Checks if graph has interfering v-structures
}
