% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCBN-density.R
\name{logLik.PCBN}
\alias{logLik.PCBN}
\title{Log-likelihood of a PCBN object}
\usage{
\method{logLik}{PCBN}(PCBN, data_uniform, ...)
}
\arguments{
\item{PCBN}{the PCBN object}

\item{data_uniform}{the dataset for which the log-likelihood is computed.
It must have already been standardized to uniform margins.}

\item{...}{other arguments, ignored for the moment}
}
\value{
the log-likelihood of the PCBN model for the given dataset
}
\description{
Log-likelihood of a PCBN object
}
\examples{
DAG = create_DAG(3)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')

order_hash = r2r::hashmap()
order_hash[['U3']] = c("U1", "U2")

fam = matrix(c(0, 1, 1,
               0, 0, 1,
               0, 0, 0), byrow = TRUE, ncol = 3)
tau = 0.2 * fam

my_PCBN = new_PCBN(
  DAG, order_hash,
  copula_mat = list(tau = tau, fam = fam))

mydata = sample_PCBN(my_PCBN, N = 10)

logLik(my_PCBN, mydata)

}
