% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCBN-object.R
\name{new_PCBN}
\alias{new_PCBN}
\title{Initializes PCBN class}
\usage{
new_PCBN(DAG, order_hash, copula_mat)
}
\arguments{
\item{DAG}{the corresponding DAG (a `bn` object)}

\item{order_hash}{a hashmap of character vectors
Each character vector corresponds
to the order of the parents for the given node.}

\item{copula_mat}{list with the matrix of families
and the matrix of estimated parameters
(assuming a 1-dimensional family)}
}
\value{
the new PCBN
}
\description{
Initializes PCBN class
}
\examples{
DAG = create_DAG(3)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')

order_hash = r2r::hashmap()
order_hash[['U3']] = c("U1", "U2")

fam = matrix(c(0, 1, 1,
               0, 0, 1,
               0, 0, 0), byrow = TRUE, ncol = 3)
tau = 0.2 * fam

my_PCBN = new_PCBN(
  DAG, order_hash,
  copula_mat = list(tau = tau, fam = fam))

}
