% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hill_climbing.R
\name{operation_do}
\alias{operation_do}
\title{Apply an operation on a DAG}
\usage{
operation_do(DAG, op)
}
\arguments{
\item{DAG}{the DAG object}

\item{op}{the operation to be applied
It should be a data frame of length 3, with a first column \code{from}, the
second \code{to}, the third \code{operation}. \code{From} and \code{to} are
the nodes (characters) and \code{operation} is a character vector describing
the operation to be applied}
}
\value{
the modified DAG. This does not modify the original DAG.
}
\description{
Apply an operation on a DAG
}
