% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaAssignement.R
\name{outgoing_arc}
\alias{outgoing_arc}
\title{Checks if w can be added by an outgoing arc}
\usage{
outgoing_arc(DAG, w, v, order_v, order_hash)
}
\arguments{
\item{DAG}{Directed Acyclic Graph.}

\item{w}{node in DAG}

\item{v}{node in DAG}

\item{order_v}{partial order for node v}

\item{order_hash}{hashmap of parental orders}
}
\value{
TRUE if w is a possible candidate, FALSE if not
}
\description{
Checks if w can be added by an outgoing arc
}
