% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{path_check}
\alias{path_check}
\title{Checks a path for converging connections and chords.}
\usage{
path_check(DAG, path)
}
\arguments{
\item{DAG}{Directed Acyclic Graph.}

\item{path}{character vector of nodes in the DAG forming a trail.}
}
\value{
\code{TRUE} if the path contains no converging connections nor chords.
}
\description{
Checks a path for converging connections and chords.
}
\examples{

DAG = create_empty_DAG(4)
DAG = bnlearn::set.arc(DAG, 'U1', 'U2')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')
DAG = bnlearn::set.arc(DAG, 'U3', 'U4')
path_check(DAG, c('U1', 'U2', 'U3', 'U4')) # TRUE

DAG = bnlearn::set.arc(DAG, 'U1', 'U4')
path_check(DAG, c('U1', 'U2', 'U3', 'U4')) # has a chord

DAG = create_empty_DAG(4)
DAG = bnlearn::set.arc(DAG, 'U1', 'U2')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')
DAG = bnlearn::set.arc(DAG, 'U4', 'U3')
path_check(DAG, c('U1', 'U2', 'U3', 'U4')) # has a converging connection

}
