% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{PCBN_sim}
\alias{PCBN_sim}
\title{Simulate data from a specified PCBN}
\usage{
PCBN_sim(object, N)
}
\arguments{
\item{object}{PCBN object to simulate from.
\bold{This does not work if the PCBN does not abide by the B-sets.
And in general, it does not work if the PCBN is outside of
the class of restricted PCBNs.}}

\item{N}{sample size}
}
\value{
a data frame of N samples
}
\description{
Simulate data from a specified PCBN
}
\examples{
DAG = create_DAG(3)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')

order_hash = r2r::hashmap()
order_hash[['U3']] = c("U1", "U2")

fam = matrix(c(0, 1, 1,
               0, 0, 1,
               0, 0, 0), byrow = TRUE, ncol = 3)

tau = 0.2 * fam

my_PCBN = new_PCBN(
  DAG, order_hash,
  copula_mat = list(tau = tau, fam = fam))

mydata = PCBN_sim(my_PCBN, N = 5)


}
