% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_activeCycles.R
\name{plot_active_cycles}
\alias{plot_active_cycles}
\title{Interactively plots a list of active cycles}
\usage{
plot_active_cycles(DAG, active_cycle_list)
}
\description{
Interactively plots a list of active cycles
}
\examples{
DAG = create_empty_DAG(5)
DAG = bnlearn::set.arc(DAG, 'U1', 'U2')
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U4')
DAG = bnlearn::set.arc(DAG, 'U3', 'U4')
DAG = bnlearn::set.arc(DAG, 'U2', 'U5')
DAG = bnlearn::set.arc(DAG, 'U3', 'U5')

activeCycles = active_cycles(DAG)  # 2 active cycles
plot_active_cycles(DAG, activeCycles)

}
