% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{sample_PCBN}
\alias{sample_PCBN}
\title{Samples from a specified PCBN}
\usage{
sample_PCBN(object, N)
}
\arguments{
\item{object}{PCBN object}

\item{N}{sample size}
}
\value{
a data frame of N samples
}
\description{
Samples from a specified PCBN
}
