% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{sym_diff}
\alias{sym_diff}
\title{Take symmetric difference of two vectors}
\usage{
sym_diff(a, b)
}
\arguments{
\item{a}{Vector 1}

\item{b}{Vector 2}
}
\value{
Symmetric difference of a and b
}
\description{
Take symmetric difference of two vectors
}
