% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{BiCopCondFit}
\alias{BiCopCondFit}
\title{Fits the copula joining w and v given cond_set abiding
by the conditional independencies of the graph}
\usage{
BiCopCondFit(
  data,
  DAG,
  w,
  v,
  cond_set,
  familyset,
  order_hash,
  margin_hash,
  copula_hash
)
}
\arguments{
\item{data}{data frame}

\item{DAG}{Directed Ayclic Graph}

\item{v, w}{nodes of the graph}

\item{cond_set}{vector of nodes of DAG in the conditioning set}

\item{familyset}{vector of copula families}

\item{order_hash}{hashmap of parental orders}

\item{margin_hash, copula_hash}{hashmaps containing already estimated objects}
}
\value{
copula object
}
\description{
Fits the copula joining w and v given cond_set abiding
by the conditional independencies of the graph
}
