% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{compute_sample_margin}
\alias{compute_sample_margin}
\title{Computes the conditional margins during sampling}
\usage{
compute_sample_margin(object, data, v, cond_set)
}
\arguments{
\item{object}{PCBN object}

\item{data}{data frame}

\item{v}{node}

\item{cond_set}{conditioning set}
}
\value{
a vector of realizations \eqn{u_{v|cond_set}}
}
\description{
Computes the conditional margins during sampling
}
