% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{ComputeCondMargin}
\alias{ComputeCondMargin}
\title{Computes the margin v given cond_set abiding by the conditional independencies of the graph}
\usage{
ComputeCondMargin(data, DAG, v, cond_set, familyset, order_hash)
}
\description{
Computes the margin v given cond_set abiding by the conditional independencies of the graph
}
