% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{ComputeCondMargin}
\alias{ComputeCondMargin}
\title{Computes the margin v given cond_set abiding by the conditional independencies of the graph}
\usage{
ComputeCondMargin(data, DAG, v, cond_set, familyset, order_hash, e, verbose)
}
\arguments{
\item{data}{data frame}

\item{DAG}{Directed Ayclic Graph}

\item{v}{node of the graph}

\item{cond_set}{vector of nodes of DAG. They should all be parents of v.
They should be ordered from the smallest to the biggest.}

\item{familyset}{vector of copula families}

\item{order_hash}{hashmap of parental orders}

\item{e}{environment containing all the hashmaps}

\item{verbose}{if \code{0}, don't print anything.
If \code{verbose = 1}, print information about the fitting procedure.}
}
\description{
Computes the margin v given cond_set abiding by the conditional independencies of the graph
}
