% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{DAG_to_restricted}
\alias{DAG_to_restricted}
\title{Turns a general graph into a restricted graph.}
\usage{
DAG_to_restricted(DAG)
}
\arguments{
\item{DAG}{Directed Acyclic Graph.}
}
\value{
Restricted DAG.
}
\description{
Turns a general graph into a restricted graph.
}
