% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{DAG_to_restricted}
\alias{DAG_to_restricted}
\title{Turns a general graph into a restricted graph.}
\usage{
DAG_to_restricted(DAG)
}
\arguments{
\item{DAG}{Directed Acyclic Graph.}
}
\value{
Restricted DAG.
}
\description{
Turns a general graph into a restricted graph.
}
\examples{
# DAG with an active cycle at node 5
DAG = create_DAG(5)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U1', 'U2')
DAG = bnlearn::set.arc(DAG, 'U2', 'U4')
DAG = bnlearn::set.arc(DAG, 'U3', 'U5')
DAG = bnlearn::set.arc(DAG, 'U4', 'U5')

# Fixed graph has extra arcs 1 -> 5, 2 -> 5
fixed_DAG = DAG_to_restricted(DAG)


# DAG with an interfering v-structures node 3
DAG = create_DAG(5)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U1', 'U4')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U5')
DAG = bnlearn::set.arc(DAG, 'U3', 'U4')
DAG = bnlearn::set.arc(DAG, 'U3', 'U5')

# Fixed graph has extra arc 1 -> 5
fixed_DAG = DAG_to_restricted(DAG)

}
