% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{dsep_set}
\alias{dsep_set}
\title{D-separation of two nodes given a set in a DAG}
\usage{
dsep_set(DAG, X, Y, Z = c())
}
\arguments{
\item{DAG}{Directed Acyclic Graph}

\item{X}{node}

\item{Y}{node}

\item{Z}{set}
}
\value{
TRUE if the sets are d-separated and FALSE if not
}
\description{
D-separation of two nodes given a set in a DAG
}
