% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimation.R
\name{fit_copulas}
\alias{fit_copulas}
\title{Fit the copulas of a PCBN given data}
\usage{
fit_copulas(
  data,
  DAG,
  order_hash,
  familyset = c(1, 3, 4, 5, 6),
  margin_hash,
  copula_hash
)
}
\arguments{
\item{data}{data frame}

\item{DAG}{Directed Acyclic Graph}

\item{order_hash}{hashmap of parental orders}

\item{familyset}{vector of copula families}

\item{margin_hash, copula_hash}{hashmaps containing already estimated objects}
}
\value{
all fitted copulas
}
\description{
Fit the copulas of a PCBN given data
}
\seealso{
[BiCopCondFit] which this function wraps.
}
