% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_B_sets.R
\name{has_interfering_vstrucs}
\alias{has_interfering_vstrucs}
\title{Checks if graph has interfering v-structures}
\usage{
has_interfering_vstrucs(DAG)
}
\arguments{
\item{DAG}{Directed Acyclic Graph}
}
\value{
TRUE if graph contains interfering vs and FALSE if not
}
\description{
Checks if graph has interfering v-structures
}
\examples{

DAG = create_DAG(5)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')

DAG = bnlearn::set.arc(DAG, 'U1', 'U4')
DAG = bnlearn::set.arc(DAG, 'U3', 'U4')
DAG = bnlearn::set.arc(DAG, 'U2', 'U5')
DAG = bnlearn::set.arc(DAG, 'U3', 'U5')

# There is one interfering v-structure
has_interfering_vstrucs(DAG)

DAG = bnlearn::set.arc(DAG, 'U1', 'U5')
# Now no interfering v-structure
has_interfering_vstrucs(DAG)

}
