% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hill.climbing.PCBN}
\alias{hill.climbing.PCBN}
\title{Hill climbing algorithm for restricted PCBNs}
\usage{
hill.climbing.PCBN(data, start, familyset, debug = FALSE)
}
\arguments{
\item{data}{data frame}

\item{start}{starting Directed Acyclic Graph}
}
\value{
DAG which locally maximizes BIC based score function
}
\description{
Hill climbing algorithm for restricted PCBNs
}
