% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hill_climbing.R
\name{hill.climbing.PCBN}
\alias{hill.climbing.PCBN}
\title{Hill climbing algorithm for restricted PCBNs}
\usage{
hill.climbing.PCBN(data, start, familyset, debug = FALSE, e)
}
\arguments{
\item{data}{data frame}

\item{start}{starting Directed Acyclic Graph}

\item{familyset}{vector of copula families}

\item{debug}{to print debug information}

\item{e}{environment containing all the hashmaps}
}
\value{
DAG which locally maximizes BIC based score function
}
\description{
Hill climbing algorithm for restricted PCBNs
}
\examples{

DAG = create_DAG(4)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U4')
DAG = bnlearn::set.arc(DAG, 'U3', 'U4')

order_hash = r2r::hashmap()
order_hash[['U3']] = c("U2", "U1")
order_hash[['U4']] = c("U2", "U3")

fam = matrix(c(0, 1, 1, 1,
               0, 0, 1, 1,
               0, 0, 0, 1,
               0, 0, 0, 0), byrow = TRUE, ncol = 4)
tau = 0.2 * fam

my_PCBN = new_PCBN(
  DAG, order_hash,
  copula_mat = list(tau = tau, fam = fam))

mydata = PCBN_sim(my_PCBN, N = 5)

# Does not work yet
# TODO: fix storing of the trees in the hash thing
#
# result = hill.climbing.PCBN(data = mydata, start = create_DAG(4),
#                             familyset = 1, debug=FALSE)

}
