% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCBN-restrictions.R
\name{is_restrictedDAG}
\alias{is_restrictedDAG}
\title{Does a DAG satisfy the restrictions of no active cycle and no
interfering v-structures}
\usage{
is_restrictedDAG(DAG, verbose = 2, check_both = TRUE)
}
\arguments{
\item{DAG}{the DAG object}

\item{verbose}{if \code{verbose} is \code{2}, details are printed.
If \code{verbose} is \code{1}, details are printed only if an active cyle
or an interfering v-structure is found.
If \code{verbose} is \code{0} the function does not print and only returns
the result.}

\item{check_both}{if \code{TRUE}, both v-structures and active cycles are
checked anyway. If \code{FALSE}, the function stops early if it already found
any v-structures.}
}
\value{
a Boolean indicating whether or not the PCBN satisfies the
restrictions.
}
\description{
Does a DAG satisfy the restrictions of no active cycle and no
interfering v-structures
}
\examples{

DAG = create_empty_DAG(4)
DAG = bnlearn::set.arc(DAG, 'U1', 'U2')
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U4')
DAG = bnlearn::set.arc(DAG, 'U3', 'U4')

is_restrictedDAG(DAG)  # 1 active cycle

}
