% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{path_check}
\alias{path_check}
\title{Checks a path for converging connections and chords.}
\usage{
path_check(DAG, path)
}
\arguments{
\item{DAG}{Directed Acyclic Graph.}

\item{path}{list of nodes in DAG forming a trail.}
}
\value{
TRUE if the path contains no converging connections and chords.
}
\description{
Checks a path for converging connections and chords.
}
