% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaAssignement.R
\name{possible_candidates}
\alias{possible_candidates}
\title{Gives possible candidates to be added to a partial order}
\usage{
possible_candidates(DAG, v, order_v, order_hash, B_minus_O)
}
\arguments{
\item{DAG}{Directed Acyclic Graph.}

\item{v}{node in DAG}

\item{order_v}{partial order for node v.}

\item{order_hash}{hashmap of parental orders}

\item{B_minus_O}{B-set setminus partial order}
}
\value{
vector of possible candidates
}
\description{
Gives possible candidates to be added to a partial order
}
