% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph.R
\name{remove_CondInd}
\alias{remove_CondInd}
\title{Remove all possible elements form conditioning set}
\usage{
remove_CondInd(DAG, node, cond_set)
}
\arguments{
\item{DAG}{Directed Acyclic Graph}

\item{node}{node}

\item{cond_set}{vector of nodes in conditioning set}
}
\value{
a vector containing the nodes not removable from the conditioning set
}
\description{
Remove all possible elements form conditioning set
}
