% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\name{sample_PCBN}
\alias{sample_PCBN}
\title{Samples from a specified PCBN}
\usage{
sample_PCBN(object, N)
}
\arguments{
\item{object}{PCBN object}

\item{N}{sample size}
}
\value{
a data frame of N samples
}
\description{
Samples from a specified PCBN
}
\examples{
DAG = create_DAG(3)
DAG = bnlearn::set.arc(DAG, 'U1', 'U3')
DAG = bnlearn::set.arc(DAG, 'U2', 'U3')

order_hash = r2r::hashmap()
order_hash[['U3']] = c("U1", "U2")

fam = matrix(c(0, 1, 1,
               0, 0, 1,
               0, 0, 0), byrow = TRUE, ncol = 3)

# FIXME
# automatically put the dimnames
# when creating PCBN objects

rownames(fam) <- c("U1", "U2", "U3")
colnames(fam) <- c("U1", "U2", "U3")
tau = 0.2 * fam

my_PCBN = new_PCBN(
  DAG, order_hash,
  copula_mat = list(tau = tau, fam = fam))

mydata = sample_PCBN(my_PCBN, N = 5)


}
